pyproject.toml
[project]
name = "task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
